+++ packages/holz-logger/src/index.node.ts
@@ -1,18 +1,23 @@
 import { isatty } from 'node:tty';
 import { AnsiTerminalBackend } from '@holz/ansi-terminal-backend';
-import { createLogger } from '@holz/core';
+import { createLogger, filter } from '@holz/core';
 import { EnvironmentFilter } from '@holz/env-filter';
 import { StreamBackend } from '@holz/stream-backend';
+
+const { NODE_ENV } = process.env;
 
 const supportsAnsiColors =
   isatty(process.stderr.fd) && process.stderr.hasColors(3);
 
 const logger = createLogger(
-  new EnvironmentFilter({
-    processor: supportsAnsiColors
-      ? new AnsiTerminalBackend()
-      : new StreamBackend({ stream: process.stderr }),
-  })
+  filter(
+    () => NODE_ENV !== 'test',
+    new EnvironmentFilter({
+      processor: supportsAnsiColors
+        ? new AnsiTerminalBackend()
+        : new StreamBackend({ stream: process.stderr }),
+    })
+  )
 );
 
 export default logger;