+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- `[@holz/logger]` Enable all server logs by default.
+
 ## [0.2.0] - 2023-02-23
 
 ### Fixed